+++ src/rule_registry.rs
@@ -5,6 +5,7 @@
 use crate::reporting::offence::Offence;
 use crate::rule::Rule;
 use crate::rules::arrange_act_assert_rule::ArrangeActAssertRule;
+use crate::rules::declared_by_name_rule::DeclaredByNameRule;
 use crate::rules::directory_file_count_rule::DirectoryFileCountRule;
 use crate::rules::directory_subfolder_count_rule::DirectorySubfolderCountRule;
 use crate::rules::header_rule::HeaderRule;
@@ -48,6 +49,7 @@ impl RuleRegistry {
         vec![
             Box::new(ReadableSourceRule::new()),
             Box::new(ArrangeActAssertRule::new()),
+            Box::new(DeclaredByNameRule::new()),
             Box::new(DirectoryFileCountRule::new(
                 config
                     .max_files_per_directory