+++ tests/rules/mod.rs
@@ -3,6 +3,7 @@
 // SPDX-License-Identifier: MIT
 
 pub mod arrange_act_assert_rule_tests;
+pub mod declared_by_name_rule_tests;
 pub mod directory_file_count_rule_tests;
 pub mod directory_subfolder_count_rule_tests;
 pub mod header_rule_tests;